form submit time for DB & timestamp consistency
